src/automation/throttling-responder/throttling-respond-py-messaging/requirements.txt
azure-functions
msrestazure==0.6.1
azure-mgmt-resource==2.2.0
azure-mgmt-cosmosdb==0.8.0